Move rotate_logs() to l.lib.logging.
